+++ CHANGELOG.md
@@ -9,3 +9,31 @@ Take the current operating system into account
 ## 0.0.17
 
 Minor improvements to the UX of the revision flow
+
+## 0.0.18
+
+Better error handling
+
+## 0.1.0
+
+Minor error message improvements
+
+## 0.1.1
+
+Minor UX improvement ("AI Shell" vs "ai-shell")
+
+## 0.1.2
+
+Improve error messaging
+
+## 0.1.3
+
+Improve error messaging, again
+
+## 0.1.4
+
+Improve error messaging, again again
+
+## 0.1.5
+
+Improve error messaging UX, again x3